+++ packages/flutter_riverpod/test/src/core/consumer_test.dart
@@ -10,7 +10,51 @@ import 'package:riverpod/legacy.dart';
 import 'package:riverpod/src/internals.dart'
     show ContainerReadElement, InternalProviderContainer;
 
+import 'provider_subscription_test.dart';
+
 void main() {
+  testWidgets('TickerMode + mutation triggers assertion', (tester) async {
+    // Regression test for https://github.com/rrousselGit/riverpod/issues/4709
+    final container = ProviderContainer.test();
+
+    final rootProvider = NotifierProvider(
+      () => DeferredNotifier((ref, self) => 0),
+    );
+    final middle = Provider<int>(
+      name: 'middle',
+      (ref) => ref.watch(rootProvider),
+    );
+    final leaf = Provider<int>(name: 'leaf', (ref) => ref.watch(middle));
+
+    final testWidget = Consumer(
+      builder: (context, ref, _) {
+        ref.watch(middle);
+        ref.watch(leaf);
+
+        return const SizedBox();
+      },
+    );
+
+    Widget app({required bool tickerModeEnabled}) {
+      return TickerMode(
+        enabled: tickerModeEnabled,
+        child: UncontrolledProviderScope(
+          container: container,
+          child: testWidget,
+        ),
+      );
+    }
+
+    // pump with TickerMode disabled
+    await tester.pumpWidget(app(tickerModeEnabled: false));
+
+    // mutate
+    container.read(rootProvider.notifier).state++;
+
+    // pump with TickerMode enabled - asserts
+    await tester.pumpWidget(app(tickerModeEnabled: true));
+  });
+
   group('Handles TickerMode', () {
     testWidgets('Does not rebuild widgets when TickerMode changes', (
       tester,