+++ packages/flutter_riverpod/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased build
+
+- Fixes assertion error when providers are unpaused.
+
 ## 3.3.2-dev.2 - 2026-05-06
 ### Dependency changes